# artisanal-kettle/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/admin/submit": {
            "post": {
                "description": "Registers a new service configuration in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Register a new service",
                "parameters": [
                    {
                        "description": "Service definition",
                        "name": "service",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.Service"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/list/services": {
            "get": {
                "description": "Returns a list of all registered service names",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "services"
                ],
                "summary": "List all service names",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error listing services",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/submit": {
            "post": {
                "description": "Submits a command to the specified service and returns the result",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "commands"
                ],
                "summary": "Submit a command to a service",
                "parameters": [
                    {
                        "description": "Command payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.payload"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.payload": {
            "type": "object",
            "properties": {
                "command": {
                    "description": "Command to execute on the service",
                    "type": "string"
                },
                "name": {
                    "description": "Name of the service to target",
                    "type": "string"
                }
            }
        },
        "service.Service": {
            "type": "object",
            "properties": {
                "isKubernetes": {
                    "description": "Whether the service is deployed on Kubernetes",
                    "type": "boolean"
                },
                "name": {
                    "description": "The name of the service",
                    "type": "string"
                },
                "server": {
                    "description": "The server address for the service",
                    "type": "string"
                }
            }
        }
    }
}